# Conflicts:
#	3rdparty/lua/Makefile
#	3rdparty/lua/README
#	3rdparty/lua/doc/contents.html
#	3rdparty/lua/doc/manual.html
#	3rdparty/lua/doc/readme.html
#	3rdparty/lua/src/ldebug.c
#	3rdparty/lua/src/ldo.c
#	3rdparty/lua/src/lua.h
#	3rdparty/portmidi/CHANGELOG.txt
#	3rdparty/portmidi/CMakeLists.txt
#	3rdparty/portmidi/Doxyfile
#	3rdparty/portmidi/README.md
#	3rdparty/portmidi/packaging/PortMidiConfig.cmake.in
#	3rdparty/portmidi/packaging/portmidi.pc.in
#	3rdparty/portmidi/pm_common/CMakeLists.txt
#	3rdparty/portmidi/pm_common/pminternal.h
#	3rdparty/portmidi/pm_common/pmutil.c
#	3rdparty/portmidi/pm_common/pmutil.h
#	3rdparty/portmidi/pm_common/portmidi.c
#	3rdparty/portmidi/pm_common/portmidi.h
#	3rdparty/portmidi/pm_haiku/pmhaiku.cpp
#	3rdparty/portmidi/pm_java/CMakeLists.txt
#	3rdparty/portmidi/pm_java/README.txt
#	3rdparty/portmidi/pm_java/pmdefaults/PmDefaults.java
#	3rdparty/portmidi/pm_java/pmdefaults/README.txt
#	3rdparty/portmidi/pm_java/pmdefaults/pmdefaults
#	3rdparty/portmidi/pm_java/pmdefaults/pmdefaults-license.txt
#	3rdparty/portmidi/pm_linux/finddefault.c
#	3rdparty/portmidi/pm_linux/pmlinux.c
#	3rdparty/portmidi/pm_linux/pmlinuxalsa.c
#	3rdparty/portmidi/pm_mac/README_MAC.txt
#	3rdparty/portmidi/pm_mac/finddefault.c
#	3rdparty/portmidi/pm_mac/pmmac.c
#	3rdparty/portmidi/pm_mac/pmmac.h
#	3rdparty/portmidi/pm_mac/pmmacosxcm.c
#	3rdparty/portmidi/pm_mac/pmmacosxcm.h
#	3rdparty/portmidi/pm_mac/readbinaryplist.c
#	3rdparty/portmidi/pm_mac/readbinaryplist.h
#	3rdparty/portmidi/pm_test/README.txt
#	3rdparty/portmidi/pm_test/fast.c
#	3rdparty/portmidi/pm_test/fastrcv.c
#	3rdparty/portmidi/pm_test/multivirtual.c
#	3rdparty/portmidi/pm_test/pmlist.c
#	3rdparty/portmidi/pm_test/recvvirtual.c
#	3rdparty/portmidi/pm_test/sendvirtual.c
#	3rdparty/portmidi/pm_test/testio.c
#	3rdparty/portmidi/pm_test/virttest.c
#	3rdparty/portmidi/pm_win/pmwin.c
#	3rdparty/portmidi/pm_win/pmwinmm.c
#	3rdparty/portmidi/porttime/porttime.h
#	3rdparty/portmidi/porttime/ptmacosx_mach.c
#	3rdparty/softfloat/mamesf.h
#	3rdparty/softfloat/milieu.h
#	3rdparty/zlib/CMakeLists.txt
#	3rdparty/zlib/ChangeLog
#	3rdparty/zlib/Makefile.in
#	3rdparty/zlib/README
#	3rdparty/zlib/configure
#	3rdparty/zlib/contrib/delphi/ZLib.pas
#	3rdparty/zlib/contrib/dotzlib/DotZLib/UnitTests.cs
#	3rdparty/zlib/contrib/infback9/inftree9.c
#	3rdparty/zlib/contrib/minizip/Makefile
#	3rdparty/zlib/contrib/minizip/configure.ac
#	3rdparty/zlib/contrib/minizip/miniunz.c
#	3rdparty/zlib/contrib/minizip/zip.c
#	3rdparty/zlib/contrib/nuget/nuget.csproj
#	3rdparty/zlib/contrib/pascal/zlibpas.pas
#	3rdparty/zlib/contrib/vstudio/readme.txt
#	3rdparty/zlib/contrib/vstudio/vc10/zlib.rc
#	3rdparty/zlib/contrib/vstudio/vc10/zlibvc.def
#	3rdparty/zlib/contrib/vstudio/vc11/zlib.rc
#	3rdparty/zlib/contrib/vstudio/vc11/zlibvc.def
#	3rdparty/zlib/contrib/vstudio/vc12/zlib.rc
#	3rdparty/zlib/contrib/vstudio/vc12/zlibvc.def
#	3rdparty/zlib/contrib/vstudio/vc14/zlib.rc
#	3rdparty/zlib/contrib/vstudio/vc14/zlibvc.def
#	3rdparty/zlib/contrib/vstudio/vc9/zlib.rc
#	3rdparty/zlib/contrib/vstudio/vc9/zlibvc.def
#	3rdparty/zlib/deflate.c
#	3rdparty/zlib/deflate.h
#	3rdparty/zlib/gzguts.h
#	3rdparty/zlib/gzlib.c
#	3rdparty/zlib/inftrees.c
#	3rdparty/zlib/old/visual-basic.txt
#	3rdparty/zlib/os400/README400
#	3rdparty/zlib/os400/zlibfixed.rpgle
#	3rdparty/zlib/qnx/package.qpg
#	3rdparty/zlib/test/minigzip.c
#	3rdparty/zlib/treebuild.xml
#	3rdparty/zlib/trees.c
#	3rdparty/zlib/win32/README-WIN32.txt
#	3rdparty/zlib/zconf.h
#	3rdparty/zlib/zconf.h.cmakein
#	3rdparty/zlib/zconf.h.in
#	3rdparty/zlib/zlib.3
#	3rdparty/zlib/zlib.3.pdf
#	3rdparty/zlib/zlib.h
#	3rdparty/zlib/zutil.h
#	scripts/src/3rdparty.lua
#	src/devices/sound/mas3507d.cpp
#	src/mame/konami/djmain.cpp
#	src/mame/konami/firebeat.cpp
#	src/mame/konami/k573dio.cpp
#	src/mame/konami/ksys573.cpp
#	src/mame/konami/twinkle.cpp
#	src/mame/sony/psx.cpp
#	src/osd/modules/font/font_sdl.cpp
#	src/osd/modules/input/input_sdl.cpp

+++ src/osd/modules/lib/osdlib_unix.cpp
@@ -2,7 +2,7 @@
 // copyright-holders:Olivier Galibert, R. Belmont
 //============================================================
 //
-//  sdlos_*.c - OS specific low level code
+//  osdlib_unix.cpp - OS specific low level code for POSIX-like systems
 //
 //  SDLMAME by Olivier Galibert and R. Belmont
 //
@@ -12,13 +12,19 @@
 #include "osdcore.h"
 #include "osdlib.h"
 
+#ifdef SDLMAME_SDL3
+#include <SDL3/SDL.h>
+#else
 #include <SDL2/SDL.h>
+#endif
 
 #include <csignal>
 #include <cstdio>
 #include <cstdlib>
+#include <cstring>
 #include <iomanip>
 #include <memory>
+#include <string_view>
 
 #include <dlfcn.h>
 #include <sys/mman.h>
@@ -60,13 +66,52 @@ void osd_process_kill()
 
 void osd_break_into_debugger(const char *message)
 {
-#ifdef MAME_DEBUG
-	printf("MAME exception: %s\n", message);
-	printf("Attempting to fall into debugger\n");
-	kill(getpid(), SIGTRAP);
+#if defined(__linux__)
+	bool do_break = false;
+	FILE *const f = std::fopen("/proc/self/status", "r");
+	if (f)
+	{
+		using namespace std::literals;
+
+		std::string_view const tag = "TracerPid:\t"sv;
+		char buf[128];
+		bool ignore = false;
+		while (std::fgets(buf, std::size(buf), f))
+		{
+			// ignore excessively long lines
+			auto const len = strnlen(buf, std::size(buf));
+			bool const noeol = !len || ('\n' != buf[len - 1]);
+			if (ignore || noeol)
+			{
+				ignore = noeol;
+				continue;
+			}
+
+			if (!std::strncmp(buf, tag.data(), tag.length()))
+			{
+				long tpid;
+				if ((std::sscanf(buf + tag.length(), "%ld", &tpid) == 1) && (0 != tpid))
+					do_break = true;
+				break;
+			}
+		}
+		std::fclose(f);
+	}
+#elif defined(MAME_DEBUG)
+	bool const do_break = true;
 #else
-	printf("Ignoring MAME exception: %s\n", message);
+	bool const do_break = false;
 #endif
+	if (do_break)
+	{
+		printf("MAME exception: %s\n", message);
+		printf("Attempting to fall into debugger\n");
+		kill(getpid(), SIGTRAP);
+	}
+	else
+	{
+		printf("Ignoring MAME exception: %s\n", message);
+	}
 }
 
 
@@ -142,7 +187,11 @@ std::error_condition osd_set_clipboard_text(std::string_view text) noexcept
 	try
 	{
 		std::string const clip(text); // need to do this to ensure there's a terminating NUL for SDL
+		#ifdef SDLMAME_SDL3
+		if (!SDL_SetClipboardText(clip.c_str()))
+		#else
 		if (0 > SDL_SetClipboardText(clip.c_str()))
+		#endif
 		{
 			// SDL_GetError returns a message, can't really convert it to an error condition
 			return std::errc::io_error; // TODO: better error code?
